Support several required fields during User creation
Custom User models may have several required attributes, for example a
username and an email. If create is called directly based on
DJANGO_USER_MAIN_ATTRIBUTE, the NOT NULL constraint will be violated,
preventing to save that user. Instead, rely on update_user to call
save when most attributes have been defined on the user.

A nice side effect is that a database query might be saved, because the
INSERT query for new users could contain all attributes when save() is
called from update_user, while a first query (INSERT) was issued to
create the user, then update user would call save(), issuing an UPDATE
query.